+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Enhancement: Enable Kotlin map-like access on CustomSamplingContext (#1192)
 * Enhancement: Auto register custom ITransportFactory in Spring integration (#1194)
 * Enhancement: Improve Kotlin property access in Performance API (#1193)
+* Enhancement: Copy options tags to transactions (#1198)
 * Enhancement: Add convenient method for accessing event's throwable (1202)
 
 # 4.0.0-alpha.3